A lot of stuff is working now.
Need to adapt the rest of the tests from FIRFilter to work with ComplexFIRFIlter

I was having this issue where the SplitComplex type's dspSCConst pointer was somehow not pointing to the right spot, the place where it was initialized. it's irking me and the hack where I made it an evaluated property is not the most elegant... getting it requires a "mutating getter" since I'm using withUnsafePointer to get the pointer to the actual dspSC member, and passing &dspSC into withUnsafePointer qualifies as mutating self. I'm going to try to ascertain what is modifying the stored property version of dspSCConst or what is going on