Packets show up in Table View, Transmit works fully
Fixed a bunch of issues in the TableViewCells, formatting issues in *CellViewModel.

Simplified Audio output API at cost of creating and destroying the Audio Output Queue every time a set of samples is played. This seems to cause much less of a headache managing buffers and state, but getting playMonoSamples(...) to work while repeatdly being spammed still has a slightly janky implementation.